% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{empirical.lh}
\alias{empirical.lh}
\title{Measure an impact of the covariates on the response using marginal empirical likelihood.}
\usage{
empirical.lh(x, y, subsamples, ...)
}
\arguments{
\item{x}{Matrix with \code{n} observations of \code{p} covariates in each row.}

\item{y}{Response vector with \code{n} observations.}

\item{subsamples}{Matrix with \code{m} indices of \code{N} subsamples in each column.}

\item{...}{Not in use.}
}
\value{
Numeric \code{p} by \code{N} matrix with empirical likelihood evaluated for each subsample.
}
\description{
This function evaluates marginal empirical likelihood between the response \code{y} and each column in the design matrix \code{x} over subsamples in \code{subsamples}.
}
\references{
Chang, Jinyuan, Cheng Yong Tang, and Yichao Wu. "Marginal empirical likelihood and sure independence feature screening." Annals of Statistics 41.4 (2013).
}
