% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{factor.model.design}
\alias{factor.model.design}
\title{Generate Factor model design matrix.}
\usage{
factor.model.design(n, p, n.factors, sigma = 1)
}
\arguments{
\item{n}{Number of independent realisations of the factor model.}

\item{p}{Number of covariates.}

\item{n.factors}{Number of factors.}

\item{sigma}{Standard deviation for the normal distribution (see details).}
}
\value{
\code{n} by \code{p} matrix with independent rows following factor model (see details).
}
\description{
Here we will add the descritpion.
}

