% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{s.est.quotient}
\alias{s.est.quotient}
\title{Estimate the size of the top-ranked set}
\usage{
s.est.quotient(prob)
}
\arguments{
\item{prob}{Vector with probabilities.}
}
\value{
A list with the following fields:
\item{scores}{Vector with the values of the criterion.}
\item{s.hat}{The estimate of the number of important covariates.}
}
\description{
Estimates the number of elements in the top-ranked set.
}
\details{
See Baranowski and Fryzlewicz (2015).
}
\references{
R. Baranowski, P. Fryzlewicz (2016), Ranking Based Variable Selection, in submission (\url{http://stats.lse.ac.uk/fryzlewicz/rbss/rbss.pdf}).
}
